improve webpack watch support
This will improve the webpack watch support (including when using
`webpack-dev-server`) in two ways:

* it will declare all bundle dependencies to webpack, so it will know
which files to watch ;

* it will use the rollup `cache` option to speed-up subsequent builds.

The cache is stored in-memory, and indexed by the global options + the
current query so it should be safe if the process is using multiple
webpack configurations at the same time.

Both changes are inspired by `rollup-watch`.